Fix tests + separate reverse proxy configuration parsing
