+++ backend/src/offspot_metrics_backend/db/initializer.py
@@ -7,7 +7,8 @@ logger = logging.getLogger(__name__)
 
 class Initializer:
     @staticmethod
-    def upgrade_db_schema(src_dir: Path):
+    def upgrade_db_schema():
         """Checks if Alembic schema has been applied to the DB"""
+        src_dir = Path(__file__).parent.parent
         logger.info(f"Upgrading database schema with config in {src_dir}")
         subprocess.check_output(args=["alembic", "upgrade", "head"], cwd=src_dir)
