--- backend/tests/unit/business/conf_missing_packages.yaml
@@ -1,1 +1,0 @@
-something: somevalue
