+++ backend/tests/unit/conf_files/conf_with_warnings.yaml
@@ -35,4 +35,3 @@ packages:
 - kind: ooo
   url: //kiwix.renaud.test/ssss
   title: toto
-
